Solved to #17 multiple source file build
