+++ examples/truestudio/Makefile.asp
@@ -165,12 +165,12 @@ CFLAGS = $(COPTS) $(CDEFS) $(INCLUDES)
 # APPL_CFG = $(APPLNAME).cfg
 
 APPL_DIR = $(APPLDIR) $(SRCDIR)/library
-APPL_ASMOBJS =
+APPL_ASMOBJS := $(APPL_ASMOBJS)
 ifdef USE_CXX
-  APPL_CXXOBJS = $(APPLNAME).o 
-  APPL_COBJS =
+  APPL_CXXOBJS := $(APPLNAME).o $(APPL_CXXOBJS)
+  APPL_COBJS := $(APPL_COBJS)
 else
-  APPL_COBJS = $(APPLNAME).o 
+  APPL_COBJS := $(APPLNAME).o $(APPL_COBJS)
 endif
 APPL_COBJS := $(APPL_COBJS) log_output.o vasyslog.o t_perror.o strerror.o
 # APPL_CFLAGS =
@@ -367,7 +367,7 @@ $(OBJNAME).srec: $(OBJFILE)
 clean:
 ifneq ($(USE_TRUESTUDIO),true)
 	rm -f $(LIB).a $(ALL_OBJ) $(DEPS)
-	rm -f \#* *~ *.o $(CLEAN_FILES)
+	rm -f \#* *~ *.o *.d $(CLEAN_FILES)
 	rm -f $(OBJFILE) $(OBJNAME).syms $(OBJNAME).srec $(OBJNAME).bin
 	rm -f kernel_cfg.timestamp $(CFG2_OUT_SRCS)
 	rm -f cfg1_out.c $(CFG1_OUT) cfg1_out.syms cfg1_out.srec
@@ -377,7 +377,7 @@ endif
 	rm -f makeoffset.s offset.h
 else
 	-rm -f $(LIB).a $(ALL_OBJ) $(DEPS)
-	-rm -f *.o *.a $(CLEAN_FILES)
+	-rm -f *.o *.d *.a $(CLEAN_FILES)
 	-rm -f $(OBJFILE) $(OBJNAME).syms $(OBJNAME).srec $(OBJNAME).bin
 	-rm -f kernel_cfg.timestamp $(CFG2_OUT_SRCS)
 	-rm -f cfg1_out.c $(CFG1_OUT) cfg1_out.syms cfg1_out.srec
